refactored SimpleAlgorithm related things to own namespace
